pipeline/requirements.txt
boto3
openpyxl
pandas
redshift_connector
python-dotenv
pymysql
cryptography
ipykernel
altair
pyarrow
awswrangler
streamlit